Add useful bits to test_helper.py
