--- tests/GalSim_LICENSE
@@ -1,22 +1,0 @@
-Copyright (c) 2012-2018 by the GalSim developers team on GitHub
-https://github.com/GalSim-developers
-
-Redistribution and use in source and binary forms, with or without
-modification, are permitted provided that the following conditions are met:
-
-1. Redistributions of source code must retain the above copyright notice, this
-   list of conditions and the following disclaimer.
-2. Redistributions in binary form must reproduce the above copyright notice,
-   this list of conditions and the following disclaimer in the documentation
-   and/or other materials provided with the distribution.
-
-This software is made available to you on an ``as is'' basis with no
-representations or warranties, express or implied, including but not
-limited to any warranty of performance, merchantability, fitness for a
-particular purpose, commercial utility, non-infringement or title.
-Neither the authors nor the organizations providing the support under
-which the work was developed will be liable to you or any third party
-with respect to any claim arising from your further development of the
-software or any products related to or derived from the software, or for
-lost profits, business interruption, or indirect special or consequential
-damages of any kind.
